requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.58.2
openai==1.3.5
weasyprint==58.1
pydyf==0.5.0
Pillow==10.1.0
python-magic==0.4.27
gunicorn==21.2.0
whitenoise==6.6.0
